acsw

+++ src/Controller/UsersController.php
@@ -2,6 +2,7 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\Event\Event;
 
 /**
  * Users Controller
@@ -20,7 +21,8 @@ class UsersController extends AppController
     public function index()
     {
         $this->paginate = [
-            'contain' => ['Belongs', 'Ranks']
+            'contain' => ['Belongs', 'Ranks', "Roles"],
+            "order" => ["user_sort_number" => "asc"]
         ];
         $users = $this->paginate($this->Users);
 
@@ -53,8 +55,6 @@ class UsersController extends AppController
         $user = $this->Users->newEntity();
         if ($this->request->is('post')) {
             $user = $this->Users->patchEntity($user, $this->request->getData());
-            $this->log("---patchEntity_user---", LOG_DEBUG);
-            $this->log($user, LOG_DEBUG);
             if ($this->Users->save($user)) {
                 $this->Flash->success(__('The user has been saved.'));
 
@@ -64,7 +64,9 @@ class UsersController extends AppController
         }
         $belongs = $this->Users->Belongs->find('list', ['limit' => 200]);
         $ranks = $this->Users->Ranks->find('list', ['limit' => 200]);
-        $this->set(compact('user', 'belongs', 'ranks'));
+        $roles = $this->authority();
+
+        $this->set(compact('user', 'belongs', 'ranks', "roles"));
     }
 
     /**
@@ -79,18 +81,26 @@ class UsersController extends AppController
         $user = $this->Users->get($id, [
             'contain' => []
         ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $user = $this->Users->patchEntity($user, $this->request->getData());
-            if ($this->Users->save($user)) {
-                $this->Flash->success(__('The user has been saved.'));
+        //認証
+        $this->Authority = $this->loadComponent("Authority");
+        if($this->Authority->userAuthorityCheck($user)){
+            if ($this->request->is(['patch', 'post', 'put'])) {
+                $user = $this->Users->patchEntity($user, $this->request->getData());
+                if ($this->Users->save($user)) {
+                    $this->Flash->success(__('The user has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('The user could not be saved. Please, try again.'));
             }
-            $this->Flash->error(__('The user could not be saved. Please, try again.'));
+        }else{
+            $this->Flash->error(__('権限がありません'));
+            return $this->redirect($this->referer());
         }
         $belongs = $this->Users->Belongs->find('list', ['limit' => 200]);
         $ranks = $this->Users->Ranks->find('list', ['limit' => 200]);
-        $this->set(compact('user', 'belongs', 'ranks'));
+        $roles = $this->authority();
+        $this->set(compact('user', 'belongs', 'ranks', "roles"));
     }
 
     /**
@@ -104,12 +114,63 @@ class UsersController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $user = $this->Users->get($id);
-        if ($this->Users->delete($user)) {
-            $this->Flash->success(__('The user has been deleted.'));
-        } else {
-            $this->Flash->error(__('The user could not be deleted. Please, try again.'));
-        }
+        $this->Authority = $this->loadComponent("Authority");
+        if($this->Authority->userAuthorityCheck($user)){
+          /*
+            //削除ではなくdelete_flagを立てる
+            $user = $this->Users->get($id, [
+                'contain' => []
+            ]);
+            $user = $this->Users->patchEntity($user, $this->request->getData());
+           */
 
+
+          /*
+            if ($this->Users->delete($user)) {
+                $this->Flash->success(__('The user has been deleted.'));
+            } else {
+                $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+            }
+           */
+        }else{
+            $this->Flash->error(__('権限がありません'));
+            return $this->redirect($this->referer());
+        }
         return $this->redirect(['action' => 'index']);
+    }
+
+    public function login()
+    {
+        if($this->request->is("post")){
+            $user = $this->Auth->identify();
+            if($user){
+                $this->Auth->setUser($user);
+                return $this->redirect($this->Auth->redirectUrl());
+            }
+            $this->Flash->error(__("ユーザ名もしくはパスワードが間違っています。"));
+        }
+        if(!empty($this->request->session()->read("Auth.User"))){
+            return $this->redirect(["controller" => "Dairy", "action" => "index"]);
+        }
+    }
+
+    public function logout()
+    {
+        return $this->redirect($this->Auth->logout());
+    }
+
+    public function authority()
+    {
+        $authority = $this->request->session()->read("Auth.User.roles_id");
+        $authority = $this->Users->Roles->find("all")
+            ->select(["role_level"])
+            ->where(["roles_id" => $authority]);
+        $authority = $authority->first()["role_level"];
+        $roles = $this->Users->Roles->find('list', [
+            'limit' => 200,
+            "order" => ["role_level" => "desc"]
+        ])
+            ->where(["role_level >=" => $authority]);
+        return $roles;
     }
 }
